+++ test/__scenarios__/zod-error.error-map.input.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod'
+
+export default function setZodLocale() {
+  const customErrorMap: z.ZodErrorMap = () => {
+    return { message: 'Custom error message' }
+  }
+
+  z.setErrorMap(customErrorMap)
+}