+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Migrate deprecated `z.setErrorMap()` to `z.config({ customError })` and `z.ZodErrorMap` to `z.core.$ZodErrorMap`. ([#65](https://github.com/nicoespeon/zod-v3-to-v4/issues/65))
+
 ## [1.11.0] - 2025-12-30
 
 ### Added